Updates to the Output Parser
Now it handles iterables in repeat blocks.
Some mappings are defined.
Cleans up some missed attributes.
Also now handles the tab character.

This code is still really dirty. It's at the point where it needs to be
refactored into a class and generally prettified. Now is not the time for
optimization, all though it could use it since almost every keyword handled
warrants another pass with the current setup.