Use a calendar widget to select dates
Currently has a bug where, in trying to open with the current date selected,
it skips a month of head but the widget doesn't increment the internal
value of the month. This resulted in a bug where Python seemed to be
claiming that there were only 30 days in October.

There's also a problem with the way the date is created that shows seconds
in the experience tree view. That should be formatted out and we should
keep the data internally.